Conclusão do CRUD de cliente pessoa. Conserto de bug ao salvar contatos, sejam eles de fornecedor, empresa ou pessoa.

+++ backend/src/main/java/br/com/dinamica/estoque/controller/ProviderContactController.java
@@ -51,7 +51,7 @@ public class ProviderContactController {
 
 	@GetMapping("/list")
 	public ResponseEntity<Object> list(
-			@RequestParam(required = false) Integer idFornecedor,
+			@RequestParam(required = false) Long idFornecedor,
 			@RequestParam(defaultValue = "0") int page,
 			@RequestParam(defaultValue = "10") int size,
 			@RequestParam(defaultValue = "id,asc") String[] sort) {
